+++ prince-archiver/tests/unit/entrypoints/test_upload_worker.py
@@ -5,7 +5,7 @@ from arq import Retry
 
 from prince_archiver.definitions import EventType, System
 from prince_archiver.entrypoints.upload_worker.worker import State, run_export
-from prince_archiver.service_layer.export import ExportHandler
+from prince_archiver.service_layer.handlers.export import ExportHandler
 from prince_archiver.service_layer.messages import ExportImagingEvent
